+++ client/components/SidebarLeft/SidebarLeft.tsx
@@ -1,10 +1,9 @@
-import React from 'react'
-import HomeIcon from '@mui/icons-material/Home'
-import CircleNotificationsIcon from '@mui/icons-material/CircleNotifications'
 import AccountCircleIcon from '@mui/icons-material/AccountCircle'
+import CircleNotificationsIcon from '@mui/icons-material/CircleNotifications'
+import HomeIcon from '@mui/icons-material/Home'
+import LogoutIcon from '@mui/icons-material/Logout'
 import SettingsIcon from '@mui/icons-material/Settings'
 import TwitterIcon from '@mui/icons-material/Twitter'
-import LogoutIcon from '@mui/icons-material/Logout';
 import { SidebarItem } from './SidebarItem'
 import SidebarTweet from './SidebarTweet'
