+++ client/components/SidebarRight/Trendbar.tsx
